Add [experimental] methods for writing arrays and dataframes (#204)
* Add methods for writing arrays and dataframes

The methods that were added to node.py and context.py are intended to be experimental. We are testing approaches to make tiled receive data and put it in a node in the tree. This might be subject to changes in the future.

* Removed deprecated method to close PR

Removed deprecated method from tifffile package that was being used for the tiff tests. A recent update of this package raised this new error

* Fixed paths

* Polish docstrings.

* Let server set mimetype.

Co-authored-by: Dan Allan <dallan@bnl.gov>