Add binary operator queries.
This adds queries for equality, binary comparison (>, <, <=, >=),
and collection membership.

Extending a pattern we use elsewhere in the code, it moves
query encoding and decoding from the client and server into
instance and class methods (respectively) on the query object
itself, putting the query objects in charge of encoding
themselves into URL query parameters.