Remove get_root_tree dependency overrides
FastAPI provides the dependency_overrides mechanism for testing and it
is not intended to be used in normal operation.

In this case, the root_tree is already set as state on the app instance
so it is accessible via fastapi's built in dependency injection using
the app field of the request.