Use pydantic to parse config files
Pydantic replaces the jsonschema validation and merge methods for the
config and provides typed access to values where the config is used.

While it paves the way for a fully typed configuration, this commit does
not make significant changes to the way the config is used after it's
parsed so for now most server settings are still passed to `build_app`
as a dictionary.